p1: added images, icons, made stuff draggable
there's a lot going on here and I haven't commented or pushed enough, so let me give you the run down of what exactly I remember doing.

So I refined the draggable images, I added in folder icons, I added in a dialogue box, I fixed the random positioning of the images, and I've added a background and custom font.