ex6: added functions , arrays and selectors
so I fetched the JSON data, and made a dataloaded function to load in the data from the JSON, but to select one element from the JSON, I made a getrandomelement function to select from one of the three arrays and return an element at random.

I also added a verb that would detect whether the condiment word ended in s or not, so it can select between the verb is or are.